+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Change Log
 
 - Fix an issue where find results was unwantedly collapsed when resizing the find panel.
 - Workaround the issue where an editor resizing required too long time.
+- Improve stability.